from kyushu

+++ surf.py
@@ -9,12 +9,13 @@ import numpy as np
 from reaction_tools import *
 
 vacuum = 10.0
-doping = False
+doping = True
 
-os.system('rm surf.db')
+if os.path.exists('surf.db'):
+	os.system('rm surf.db')
 
-nlayer = 5
-nrelax = 3
+nlayer = 2
+nrelax = 1
 
 cif_file = "mgo.cif"
 #cif_file = "pd.cif"
@@ -23,7 +24,7 @@ cif_file = "mgo.cif"
 #cif_file = "Ce2W3O12.cif"
 
 lattice = "fcc"
-facet   = "310"
+facet   = "100"
 #lattice = "hcp"
 #facet   = "001"
 #lattice = "sp15"
@@ -40,22 +41,22 @@ if cif_file == "La2O3.cif":
 	surf.wrap()
 	surf.center(axis=2) # La2O3, only z-axis
 
-surf = surf*[1,2,1]
+surf = surf*[2,2,1]
 #surf = sort(surf)
 surf = sort_atoms_by_z(surf)
 
 formula = surf.get_chemical_formula()
 
-#offset_fac = 3.0/4.0
+offset_fac = 3.0/4.0
 #offset_fac = 1.5 # for Pd110*[2,2,1]
 #offset_fac = 1.0
-offset_fac = 0.67
+#offset_fac = 0.67
 #
 # doping e.g.) Mg by Li
 #
 if doping:
 	symbols  =  np.array(surf.get_chemical_symbols())
-	rep_atom = 112
+	rep_atom = 48
 	symbols[rep_atom] = 'Li'
 	surf.set_chemical_symbols(symbols)
 
@@ -90,4 +91,6 @@ surf.set_tags(tag)
 
 db = connect("surf.db")
 db.write(surf, data=pos)
+
+view(surf)
